Extract getBase64Creds to a shared module
