+++ parodos-model-api/src/main/java/com/redhat/parodos/workflow/utils/CredUtils.java
@@ -1,0 +1,21 @@
+package com.redhat.parodos.workflow.utils;
+
+import java.util.Base64;
+
+public class CredUtils {
+
+	/**
+	 * Generates a Base64 encoding of username:password string
+	 * @param username the username
+	 * @param password the password
+	 * @return the Base64 encoded string
+	 */
+	public static String getBase64Creds(String username, String password) {
+		String plainCreds = username + ":" + password;
+		byte[] plainCredsBytes = plainCreds.getBytes();
+		byte[] base64CredsBytes = Base64.getEncoder().encode(plainCredsBytes);
+		String base64Creds = new String(base64CredsBytes);
+		return base64Creds;
+	}
+
+}
