+++ parodos-model-api/src/test/java/com/redhat/parodos/workflow/utils/CredUtilsTest.java
@@ -1,0 +1,20 @@
+package com.redhat.parodos.workflow.utils;
+
+import org.junit.Test;
+
+import static org.junit.Assert.assertEquals;
+
+public class CredUtilsTest {
+
+	@Test
+	public void testGetBase64Creds() {
+		String username = "user123";
+		String password = "pass456";
+		String expectedBase64Creds = "dXNlcjEyMzpwYXNzNDU2";
+
+		String actualBase64Creds = CredUtils.getBase64Creds(username, password);
+
+		assertEquals(expectedBase64Creds, actualBase64Creds);
+	}
+
+}
