+++ workflow-examples/src/main/java/com/redhat/parodos/examples/utils/RestUtils.java
@@ -1,5 +1,6 @@
 package com.redhat.parodos.examples.utils;
 
+import com.redhat.parodos.workflow.utils.CredUtils;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
@@ -7,8 +8,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.client.RestTemplate;
 
 import java.net.URI;
-import java.util.Base64;
-import java.util.Map;
 
 /**
  * RestUtils is an utility class. All its methods must be declared as static so they can't
@@ -63,24 +62,10 @@ public final class RestUtils {
 	 * @return the @see org.springframework.http.HttpEntity
 	 */
 	public static HttpEntity<String> getRequestWithHeaders(String username, String password) {
-		String base64Creds = getBase64Creds(username, password);
+		String base64Creds = CredUtils.getBase64Creds(username, password);
 		HttpHeaders headers = new HttpHeaders();
 		headers.add("Authorization", "Basic " + base64Creds);
 		return new HttpEntity<>(headers);
-	}
-
-	/**
-	 * Generates a Base64 encoding of username:password string
-	 * @param username the username
-	 * @param password the password
-	 * @return the Base64 encoded string
-	 */
-	public static String getBase64Creds(String username, String password) {
-		String plainCreds = username + ":" + password;
-		byte[] plainCredsBytes = plainCreds.getBytes();
-		byte[] base64CredsBytes = Base64.getEncoder().encode(plainCredsBytes);
-		String base64Creds = new String(base64CredsBytes);
-		return base64Creds;
 	}
 
 }
